Calculate height from font metrics
This is an alternate solution to the previous one, where the font
metrics are used to determine the image height. This has the benefit
that it will always have the correct height, regardless of the character
range to import (whereas the other one would adapt to the ascenders and
descenders of the character range).